# openapi/blerp.v1.yaml
openapi: 3.1.0
info:
  title: Blerp Identity Service API
  version: 1.0.0-draft
  contact:
    name: Blerp Support
    url: https://github.com/e6qu/blerp
    email: adi11235@gmail.com
  summary: REST + JSON interface for Blerp Identity Service
  description: >
    Clean-room OpenAPI contract describing the public REST API for user, session,
    organization, and webhook management. This specification is the authoritative
    contract for backend/frontend work and must exist locally before any detailed
    implementation plan is executed.
...
servers:
  - url: https://api.blerp.dev
    description: Production
  - url: https://staging-api.blerp.dev
    description: Staging
  - url: http://localhost:4000
    description: Local development (docker-compose)
security:
  - SecretKey: []
...
components:
  securitySchemes:
    SecretKey:
      type: http
      scheme: bearer
      bearerFormat: secret-key
      description: Server-to-server secret API key (sk_live_* / sk_test_*).
    PublishableKey:
      type: http
      scheme: bearer
      bearerFormat: publishable-key
      description: Client-side publishable API key (pk_live_* / pk_test_*).
    SessionToken:
      type: http
      scheme: bearer
      bearerFormat: session-token
      description: Session token returned after sign-in/up (`sess_*`).
  schemas:
    ErrorResponse:
      type: object
      # BUG-47/57: blerp emits the Clerk-compatible `errors[]` envelope
      # (canonical for new SDK code) from any path that throws a
      # `BlerpError`. Many hand-rolled `res.status(N).json({ error: ... })`
      # paths in controller catch blocks still emit only the singular
      # `error` field — so `error` is required while `errors` stays
      # optional until those sites migrate to throwing `BlerpError`
      # (tracked in DO_NEXT.md). Generated SDK clients should read
      # `errors[0]` when present and fall back to `error` otherwise.
      # BUG-60 (codex r4): had `errors` required here; SDK consumers
      # would crash on legacy-shape 400s.
      required: [error]
      properties:
        errors:
          type: array
          minItems: 1
          items:
            type: object
            required: [code, message, long_message]
            properties:
              code:
                type: string
              message:
                type: string
              long_message:
                type: string
              meta:
                type: object
                additionalProperties: true
        error:
          type: object
          required: [code, message]
          description: Legacy singular alias; prefer `errors[0]` in new code.
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            doc_url:
              type: string
              format: uri
            request_id:
              type: string
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          minimum: 0
        next_cursor:
          type: string
        previous_cursor:
          type: string
...
    Membership:
      type: object
      required: [id, organization_id, user_id, role]
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        invited_by:
          type: string
          format: uuid
        status:
          type: string
          enum: [active, revoked, pending_invite]
        last_active_at:
          type: string
          format: date-time
    Invitation:
      type: object
      required:
        - id
        - email
        - organization_id
        - role
        - status
        - created_at
        - expires_at
      properties:
        id:
          type: string
        project_id:
          type: string
        organization_id:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
        permissions:
          type: array
          items:
            type: string
        inviter_user_id:
          type: string
        status:
          type: string
          enum: [pending, accepted, revoked, expired]
        redirect_url:
          type: string
          format: uri
        created_at:
          type: string
          format: date-time
        expires_at:
          type: string
          format: date-time
        accepted_at:
          type: string
          format: date-time
...
paths:
...
  /v1/organizations/{organization_id}/memberships:
    get:
      tags: [Organizations]
      summary: List memberships
      description: Returns a list of memberships for the specified organization.
      operationId: listMemberships
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of memberships
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Membership"
    post:
      tags: [Organizations]
      summary: Create membership
      description: Adds a user to an organization with the provided roles.
      operationId: createMembership
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, role]
              properties:
                user_id:
                  type: string
                role:
                  type: string
      responses:
        "201":
          description: Membership created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
  /v1/organizations/{organization_id}/memberships/me:
    get:
      tags: [Organizations]
      summary: Get the caller's own membership in this organization
      description: |
        Returns the calling user's membership row plus the authoritative
        resolved `permissions` array. Gated only by `authMiddleware` —
        every authenticated user can ask about THEIR OWN role +
        permissions without needing `members:read`, which custom
        read-only roles may lack. This is the endpoint
        `@blerp/nextjs auth()` uses to populate `orgPermissions` for
        `has({ permission })` checks.
      operationId: getOwnMembership
      # BUG-75 (codex r11): authMiddleware accepts the user's session
      # JWT, not an sk_* secret key. Documented as SessionToken so
      # generated clients send the right credential.
      security:
        - SessionToken: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: The caller's membership in the target organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "404":
          description: Caller is not a member of this organization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/organizations/{organization_id}/memberships/{membership_id}:
    patch:
      tags: [Organizations]
      summary: Update membership
      description: Updates a membership's role or permissions.
      operationId: updateMembership
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
        - name: membership_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                role:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Membership"
        "404":
          description: Not found
    delete:
      tags: [Organizations]
      summary: Delete membership
      description: Removes a user from an organization.
      operationId: deleteMembership
      security:
        - SecretKey: []
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
        - name: membership_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Membership deleted
        "404":
          description: Not found
...
  /v1/invitations:
    get:
      tags: [Invitations]
      summary: List invitations
      description: Lists project invitations including pending, accepted, and expired states.
      operationId: listInvitations
      security:
        - SecretKey: []
      parameters:
        - in: query
          name: email
          schema:
            type: string
            format: email
        - in: query
          name: organization_id
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, accepted, revoked, expired]
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Invitations
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Invitation"
                  meta:
                    $ref: "#/components/schemas/PaginationMeta"
    post:
      tags: [Invitations]
      summary: Create invitation
      description: Creates a new invitation for a user to join the project or organization.
      operationId: createInvitation
      security:
        - SecretKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, organization_id, role]
              properties:
                email:
                  type: string
                  format: email
                organization_id:
                  type: string
                role:
                  type: string
                permissions:
                  type: array
                  items:
                    type: string
                expires_at:
                  type: string
                  format: date-time
                redirect_url:
                  type: string
                  format: uri
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
  /v1/invitations/{invitation_id}/revoke:
    post:
      tags: [Invitations]
      summary: Revoke invitation
      description: Revokes an existing invitation so it can no longer be used.
      operationId: revokeInvitation
      security:
        - SecretKey: []
      parameters:
        - in: path
          name: invitation_id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        "200":
          description: Invitation revoked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invitation"
        "404":
          description: Not found
...